Add basic stat info on week to week
