+++ PyFL/constants.py
@@ -1,0 +1,9 @@
+baseurl = 'http://api.fantasy.nfl.com/v1/players/stats?statType=weekStats&season=2015&format=json&week='
+
+valid_positions = (
+    'QB',
+    'RB',
+    'WR',
+    'TE',
+    'DEF'
+)
